move echelle updating to pair and register programs, remove update_level forever

+++ GoatLib.pm
@@ -3,6 +3,7 @@ package GoatLib;
 use strict;
 use Pod::Usage;
 
+use Encode;
 use Exporter;
 use Time::ParseDate;  # This is in Debian's libtime-modules-perl package
 use Date::Parse;
@@ -15,7 +16,7 @@ use GoatConfig;
 
 our @ISA=qw(Exporter);
 our @EXPORT=qw( stone_to_level level_to_stones download_echelle parse_datestr 
-utc2str my_time
+utc2str my_time update_ech
 );
 
 =head2 fixup_frenchisms
@@ -127,10 +128,79 @@ sub level_to_stones {
 sub download_echelle {
     my ($filename) = @_;
 
+    return 1 if defined $CFG->{testing};
+
     unlink $filename;
     `wget http://ffg.jeudego.org/echelle/echtxt/ech_ffg_new.txt`;
     return -r $filename;
 }
+
+
+# Update players' level, name, clubs, license, ...
+sub update_fields {
+    my ($tournament, %all_players) = @_;
+    foreach my $player ($tournament->players) {
+        if (not exists $all_players{$player->id}) {
+            warn $player->fullname . " not found in echelle file -- skipping\n";
+            next;
+        }
+
+        my $ech_p = $all_players{$player->id};
+
+        if ($ech_p->fullname ne $player->fullname) {
+            warn "Updating name from ".$player->fullname." to ".$ech_p->fullname."\n";
+            $player->givenname($ech_p->givenname);
+            $player->familyname($ech_p->familyname);
+        }
+        if ($ech_p->level != $player->level) {
+            warn $player->fullname.": changing from ".$player->level." to ".$ech_p->level."\n";
+            $player->level($ech_p->level);
+
+        }
+        if ($ech_p->club ne $player->club) {
+            warn $player->fullname.": updating club from ".$player->club." to ".$ech_p->club."\n";
+            $player->club($ech_p->club);
+        }
+        if ($ech_p->license ne $player->license) {
+            warn $player->fullname.": updating license from ".$player->license." to ".$ech_p->license."\n";
+            $player->license($ech_p->license);
+        }
+        if ($ech_p->status ne $player->status) {
+            warn $player->fullname.": license status changed to:". $ech_p->status."\n";
+            $player->status($ech_p->status);
+        }
+        if (not $player->is_licensed($TOURNAMENT_LICENSES)) {
+            warn $player->fullname.": player is not licensed.\n";
+            warn $player->fullname.": player license: ".$player->status." is not allowed.\n"; 
+        }
+    }
+}
+
+# Download a new echelle and update all player fields if there is a tournament
+sub update_ech {
+    my ($tournament) = @_;
+
+    # Build a hash of all the players in the echelle file
+    my %all_players;
+    my $ech;
+
+    my $ECHELLE_FILE = "ech_ffg_new.txt";
+    my $r = download_echelle($ECHELLE_FILE);
+    return $r if not defined $tournament;
+
+    open $ech, $ECHELLE_FILE or die "$ECHELLE_FILE: $!\n";
+
+
+    while (<$ech>) {
+        next if /^#/;
+        my $p = new_from_ech FFGPlayer decode 'ISO-8859-1', $_;
+        next if not defined $p;
+        $all_players{$p->id} = $p;
+    }
+
+    update_fields($tournament, %all_players);
+}
+
 
 
 # Turn a UTC time_t into a localised timezoned string
